+++ docutils/HISTORY.rst
@@ -14,8 +14,14 @@
 .. contents::
 
 
-Release 0.22rc2 (unpublished)
+Release 0.22rc3 (unpublished)
 =============================
+
+* nothing yet.
+
+Release 0.22rc2 (2025-05-22)
+============================
+
 
 * docutils/parsers/rst/directives/misc.py